Removes the CONST_create_template folder only if he exists

+++ c2cgeoportal/scripts/c2ctool.py
@@ -346,7 +346,8 @@ class C2cTool:
                 pip_cmd += ["c2cgeoportal=={0!s}".format((self.options.version))]
             check_call(pip_cmd)
 
-        shutil.rmtree("CONST_create_template")
+        if os.path.exists("CONST_create_template"):
+            shutil.rmtree("CONST_create_template")
         check_call([
             "{0!s}/pcreate".format(self.venv_bin), "--ignore-conflicting-name", "--overwrite",
             "--scaffold=c2cgeoportal_update", "../{0!s}".format(self.project["project_folder"])
